Adding new role Users and making a new Departmets seeder

+++ SiT Access Control System/Services/SACS.Services.Data/DepartmentService.cs
@@ -26,14 +26,31 @@
 
         public Department GetDepartmentById(string id)
         {
-            return this.departmentRepository.All().FirstOrDefault(x => x.Id == id);
+            int departmentId;
+            if (int.TryParse(id, out departmentId))
+            {
+                return this.departmentRepository.All().FirstOrDefault(x => x.Id == departmentId);
+            }
+            return null;
         }
 
         public void RemoveById(string id)
         {
-            Department choosenDepartment = this.departmentRepository.All().FirstOrDefault(x => x.Id == id);
-            this.departmentRepository.Delete(choosenDepartment);
-            this.departmentRepository.SaveChangesAsync();
+            int departmentId;
+            if (int.TryParse(id, out departmentId))
+            {
+                Department choosenDepartment = this.departmentRepository.All().FirstOrDefault(x => x.Id == departmentId);
+                if (choosenDepartment != null)
+                {
+                    this.departmentRepository.Delete(choosenDepartment);
+                    this.departmentRepository.SaveChangesAsync();
+                }
+            }
+            else
+            {
+                // Handle the case where the id is not a valid integer, if needed
+                throw new ArgumentException("Invalid department ID");
+            }
         }
 
         public List<Department> GetAll()
